Move expand path tests to seperate test module
