+++ astrality/tests/config/test_config.py
@@ -10,8 +10,6 @@ from astrality.config import (
     create_config_directory,
     dict_from_config_file,
     user_configuration,
-    expand_path,
-    expand_globbed_path,
     insert_into,
     resolve_config_directory,
 )
@@ -154,46 +152,6 @@ class TestCreateConfigDirectory:
         assert 'astrality.yml' in dir_contents
         assert 'modules' in dir_contents
         rmtree(created_config_dir)
-
-def test_expand_path_method(test_config_directory):
-    absolute_path = Path('/tmp/ast')
-    tilde_path = Path('~/dir')
-    relative_path = Path('test')
-
-    assert expand_path(
-        path=absolute_path,
-        config_directory=Path('/what/ever'),
-    ) == absolute_path
-
-    assert expand_path(
-        path=tilde_path,
-        config_directory=Path('/what/ever'),
-    ) == Path.home() / 'dir'
-
-    assert expand_path(
-        path=relative_path,
-        config_directory=test_config_directory,
-    ) == test_config_directory / 'test'
-
-def test_expand_globbed_path(test_config_directory):
-    """Globbed paths should allow one level of globbing."""
-    templates = Path('test_modules', 'using_all_actions')
-    paths = expand_globbed_path(
-        path=templates / '*',
-        config_directory=test_config_directory,
-    )
-    assert len(paths) == 5
-    assert test_config_directory / templates / 'module.template' in paths
-
-def test_expand_recursive_globbed_path(test_config_directory):
-    """Globbed paths should allow recursive globbing."""
-    templates = Path('test_modules', 'using_all_actions')
-    paths = expand_globbed_path(
-        path=templates / '**' / '*',
-        config_directory=test_config_directory,
-    )
-    assert len(paths) == 6
-    assert test_config_directory / templates / 'recursive' / 'empty.template' in paths
 
 @pytest.yield_fixture
 def dir_with_compilable_files(tmpdir):
